==> code/02_image_stitching/06-spatial_coords.R <==
library(here)
library(tidyverse)
library(jaffelab)
library(rjson)
library(cowplot)
library(grid)
library(gridExtra)
library(getopt)
library(sessioninfo)
library(Polychrome)
data(palette36)

spec <- matrix(
    c("donor", "d", 1, "character", "Donor"),
    byrow = TRUE, ncol = 5
)
opt <- getopt(spec)

#   Column names for raw tissue_positions_list.csv files
tissue_colnames <- c(
    "barcode", "in_tissue", "array_row", "array_col", "pxl_row_in_fullres",
    "pxl_col_in_fullres"
)

tissue_path_in <- here(
    "processed-data", "02_image_stitching",
    sprintf("tissue_positions_%s.csv", opt$donor)
)

tissue_path_out <- here(
    "processed-data", "04_VisiumStitcher", opt$donor,
    "tissue_positions.csv"
)

sample_info_path <- here(
    "processed-data", "02_image_stitching", "sample_info_clean.csv"
)

plot_dir <- here("plots", "02_image_stitching", opt$donor)

dir.create(plot_dir, showWarnings = FALSE)

#   55-micrometer diameter for Visium spot; 100 micrometers between spots;
#   65-micrometer spot diameter used in 'spot_diameter_fullres' calculation for
#   spaceranger JSON. See documentation for respective quantities. The
#   difference between 55 and 65 does indeed exist and is properly documented,
#   but is likely a bug in the sense the choice was probably unintentional
#   https://kb.10xgenomics.com/hc/en-us/articles/360035487812-What-is-the-size-of-the-spots-on-the-Visium-Gene-Expression-Slide-
#   https://kb.10xgenomics.com/hc/en-us/articles/360035487892-How-much-space-is-there-between-spots-referred-to-as-white-space-
#   https://support.10xgenomics.com/spatial-gene-expression/software/pipelines/latest/output/spatial
SPOT_DIAMETER_M <- 55e-6
SPOT_DIAMETER_JSON_M <- 65e-6
INTER_SPOT_DIST_M <- 100e-6

set.seed(0)

################################################################################
#   Functions
################################################################################

#   Stop if the tibble-like 'coords', expected to contain columns 'array_row'
#   and 'array_col', represents an invalid Visium array
validate_array <- function(coords) {
    #   Even array rows can only use even column indices
    coords |>
        filter(array_row %% 2 == 0) |>
        summarize(a = all(array_col %% 2 == 0)) |>
        pull(a) |>
        stopifnot()

    #   Odd array rows can only use odd column indices
    coords |>
        filter(array_row %% 2 == 1) |>
        summarize(a = all(array_col %% 2 == 1)) |>
        pull(a) |>
        stopifnot()

    #   Check lower bound of array row and col (note we're allowing arbitrary
    #   maximum values rather than the convention of 78 rows and 128 columns)
    stopifnot(min(coords$array_row) %in% c(0, 1))
    stopifnot(min(coords$array_col) %in% c(0, 1))

    #   Check an eccentric detail of Visium arrays: (0, 0) cannot exist
    if (any((coords$array_row == 0) & (coords$array_col == 0))) {
        stop("The invalid array coordinate (0, 0) exists after fitting")
    }
}

refine_fit <- function(x, y, INTERVAL_X, INTERVAL_Y) {
    #   Round x to the nearest integer, and track the error from doing so in the
    #   variable 'dx'
    dx <- x - clean_round(x)
    x <- x - dx

    #   Given x, round y to the nearest valid integer (y must be even iff x is),
    #   and track the error from doing so in the variable 'dy'
    dy <- rep(0, length(y))
    dy[x %% 2 == 0] <- y[x %% 2 == 0] - clean_round(y[x %% 2 == 0] / 2) * 2
    dy[x %% 2 == 1] <- y[x %% 2 == 1] - (clean_round(y[x %% 2 == 1] / 2 - 0.5) * 2 + 1)
    y <- y - dy

    #   Summarize error in Euclidean distance
    error <- sqrt((INTERVAL_X * dx)**2 + (INTERVAL_Y * dy)**2)
    return(list(x, y, error))
}

#   Given 'coords', a tibble whose rows represent samples of the same
#   donor, and containing columns 'array_row', 'array_col',
#   'pxl_row_in_fullres', and 'pxl_col_in_fullres', modify the 'array_row' and
#   'array_col' columns to represent a larger Visium capture area containing
#   all samples in a common coordinate system. 'inter_spot_dist_px' gives the
#   distance between any 2 spots in the new coordinates, and so the number of
#   array rows/cols generally changes from the Visium standards of 78 and 128
#   (and even changes in ratio between num rows and num cols). Return the same
#   tibble with modified 'array_row' + 'array_col' columns, as well as new
#   'pxl_row_rounded' and 'pxl_col_rounded' columns representing the pixel
#   coordinates rounded to the nearest exact array coordinates.
fit_to_array <- function(coords, inter_spot_dist_px) {
    MIN_ROW <- min(coords$pxl_col_in_fullres)
    INTERVAL_ROW <- inter_spot_dist_px * cos(pi / 6)

    MIN_COL <- min(coords$pxl_row_in_fullres)
    MAX_COL <- max(coords$pxl_row_in_fullres)
    INTERVAL_COL <- inter_spot_dist_px / 2

    #   Calculate what "ideal" array rows and cols should be (allowed to be any
    #   float). Don't round yet. Note array_row maps with pxl_col, while
    #   array_col maps backwards with pxl_row
    array_row_temp <- (coords$pxl_col_in_fullres - MIN_ROW) /
        INTERVAL_ROW

    array_col_temp <- (MAX_COL - coords$pxl_row_in_fullres) /
        INTERVAL_COL

    #   For now, find the nearest row first, then round to the nearest possible
    #   column given the row
    temp <- refine_fit(array_row_temp, array_col_temp, INTERVAL_ROW, INTERVAL_COL)
    error_row_first <- temp[[3]]
    coords$array_row <- temp[[1]]
    coords$array_col <- temp[[2]]

    #   Perform the opposite order (column then row). When this ordering results
    #   in lower error, use it instead
    temp <- refine_fit(array_col_temp, array_row_temp, INTERVAL_COL, INTERVAL_ROW)
    error_col_first <- temp[[3]]
    coords$array_row[error_row_first > error_col_first] <- temp[[2]][
        error_row_first > error_col_first
    ]
    coords$array_col[error_row_first > error_col_first] <- temp[[1]][
        error_row_first > error_col_first
    ]

    #   Now make new pixel columns based on just the array values (these columns
    #   give the coordinates for given array row/cols)
    coords$pxl_col_rounded <- MIN_ROW + coords$array_row * INTERVAL_ROW
    coords$pxl_row_rounded <- MAX_COL - coords$array_col * INTERVAL_COL

    #-------------------------------------------------------------------------------
    #   array (0, 0) does not exist on an ordinary Visium array. Move any such
    #   values to the nearest alternatives
    #-------------------------------------------------------------------------------

    #   Nearest points to (0, 0) are (0, 2) and (1, 1):
    array_02 <- c(MIN_ROW, MIN_COL + 2 * INTERVAL_COL)
    array_11 <- c(MIN_ROW + INTERVAL_ROW, MIN_COL + INTERVAL_COL)

    #   Determine the distances to those nearest points
    dist_coords <- coords |>
        filter(array_row == 0, array_col == 0) |>
        mutate(
            dist_02 = (pxl_col_in_fullres - array_02[1])**2 +
                (pxl_row_in_fullres - array_02[2])**2,
            dist_11 = (pxl_col_in_fullres - array_11[1])**2 +
                (pxl_row_in_fullres - array_11[2])**2,
        )

    #   Move any instances of (0, 0) to the nearest alternative
    indices <- (coords$array_row == 0) & (coords$array_col == 0)
    coords[indices, "array_row"] <- ifelse(
        dist_coords$dist_02 < dist_coords$dist_11, 0, 1
    )
    coords[indices, "array_col"] <- ifelse(
        dist_coords$dist_02 < dist_coords$dist_11, 2, 1
    )

    #   Verify the newly assigned array row and cols have reasonable values
    validate_array(coords)

    return(coords)
}

#   Round to the nearest integer, always rounding up at 0.5. This consistent
#   behavior is favorable for our application, where we want to minimize
#   duplicate mappings
clean_round <- function(x) {
    return(floor(x) + ((x * 10) %% 10 >= 5))
}

################################################################################
#   Read in sample info and spot coordinate info
################################################################################

coords <- read.csv(tissue_path_in) |>
    as_tibble() |>
    mutate(sample_id = paste(ss(key, "_", 2), ss(key, "_", 3), sep = "_"))

sample_info <- read.csv(sample_info_path) |>
    as_tibble() |>
    rename(
        sample_id = X,
        array_num = "Array..",
        slide_num = "Slide.."
    ) |>
    filter(Brain == opt$donor, In.analysis == "True")

#-------------------------------------------------------------------------------
#   Read in the untransformed and unfiltered (by "in tissue") spot coordinates
#   to verify the relationship between array and pixel coordinates
#-------------------------------------------------------------------------------

raw_tissue_paths <- sample_info |>
    pull(spaceranger_dir) |>
    list.files(pattern = "^tissue_positions(_list|)\\.csv$", full.names = TRUE)

#   Read in coordinates, keeping track of sample ID as well
raw_tissue_list <- list()
for (raw_tissue_path in raw_tissue_paths) {
    raw_tissue_list[[raw_tissue_path]] <- read.csv(raw_tissue_path, col.names = tissue_colnames)
    raw_tissue_list[[raw_tissue_path]]$sample_id <- raw_tissue_path |>
        str_extract("/[^/]*_[ABCD]1/") |>
        str_replace_all("/", "")
}
raw_coords <- do.call(rbind, raw_tissue_list) |> as_tibble()

print("Correlation relationship between array (a) and pixel (p) row/col:")
cor_raw_coords <- raw_coords |>
    group_by(sample_id) |>
    summarize(
        arow_v_prow = cor(array_row, pxl_row_in_fullres),
        arow_v_pcol = cor(array_row, pxl_col_in_fullres),
        acol_v_prow = cor(array_col, pxl_row_in_fullres),
        acol_v_pcol = cor(array_col, pxl_col_in_fullres)
    )
print(cor_raw_coords)

#   We expect array_row to "line up with" pixel_col
tol <- 0.01
stopifnot(all(abs(cor_raw_coords$arow_v_prow) < tol))
stopifnot(all(1 - abs(cor_raw_coords$arow_v_pcol) < tol))

#-------------------------------------------------------------------------------
#   Select an appropriate number of array rows and columns for this slide such
#   that the distance between spots on the new grid matches the Visium standard
#-------------------------------------------------------------------------------

sr_json <- fromJSON(
    file = file.path(
        sample_info$spaceranger_dir[1], "scalefactors_json.json"
    )
)

#   Since we have the spot diameter both in pixels and meters, compute the
#   image's pixel/m ratio. Then use that to compute the distance between spots
#   in pixels
PX_PER_M <- sr_json$spot_diameter_fullres / SPOT_DIAMETER_JSON_M
INTER_SPOT_DIST_PX <- INTER_SPOT_DIST_M * PX_PER_M

#   Sanity check on existing coordinates: spots should be 100um apart. TODO:
#   look for pairs of array coordinates that exist; don't hardcode 10, 11
a <- coords |>
    filter(array_row == 10, array_col == 80) |>
    head(1) |>
    select(pxl_row_in_fullres, pxl_col_in_fullres)

b <- coords |>
    filter(array_row == 11, array_col == 81) |>
    head(1) |>
    select(pxl_row_in_fullres, pxl_col_in_fullres)

tol <- 2
observed_dist <- sqrt(rowSums((a - b)**2))
if (abs(observed_dist - INTER_SPOT_DIST_PX) > tol) {
    stop("Transformed spots output from ImageJ/Samui were not close to 100um apart!")
}

################################################################################
#   Visually assess array_row and array_col after fitting to "new grid"
################################################################################

#   Adjust 'array_row' and 'array_col' with values appropriate for the new
#   coordinate system (a larger Visium grid with equal inter-spot distances)
coords <- fit_to_array(coords, INTER_SPOT_DIST_PX)

coords |>
    #   Make barcode compatible with ImageJ outputs
    mutate(key = paste0(sample_id, ss(key, "_"))) |>
    rename(barcode = key) |>
    #   Drop columns not in spaceranger output
    select(!c(pxl_row_rounded, pxl_col_rounded, sample_id)) |>
    #   Fix data type of some columns
    mutate_at(
        c("array_row", "array_col", "pxl_row_in_fullres", "pxl_col_in_fullres"),
        ~ as.integer(round(.))
    ) |>
    write.csv(file = tissue_path_out, row.names = FALSE, quote = FALSE)

#-------------------------------------------------------------------------------
#   Spot plots
#-------------------------------------------------------------------------------

#   Plot the transformed spots as-is for reference
p <- ggplot(coords) +
    geom_point(
        aes(
            x = pxl_col_in_fullres,
            y = max(pxl_row_in_fullres) - pxl_row_in_fullres,
            color = sample_id
        ),
        size = 0.1
    ) +
    coord_fixed() +
    labs(
        x = "Full-res pixels (width)", y = "Full-res pixels (height)",
        color = "Sample ID"
    ) +
    guides(color = guide_legend(override.aes = list(size = 1.5)))

#   'Br8492' will be used for a manuscript figure, where we have stricter
#   aesthetic preferences tailored to this donor
if (opt$donor == "Br8492") {
    sample_colors = c(palette36[c(7, 6)], "#45200D", palette36[8])
    names(sample_colors) = unique(sample_info$sample_id)
    p = p + scale_color_manual(values = sample_colors)
}

pdf(file.path(plot_dir, "raw_spots.pdf"))
print(p)
dev.off()

#   To avoid overplotting, we'll scale alpha to always be 1, no matter how many
#   spots belong to a particular pair of array coordinates. This also makes
#   the overlap of spots look visually intuitive
coords_plotting = coords |>
    #   Drop duplicated array coordinates within each sample
    group_by(sample_id, array_row, array_col) |>
    filter(row_number() == 1) |>
    #   Scale alpha such that the sum of alphas across spots at a unique
    #   array location is always 1
    group_by(array_row, array_col) |>
    mutate(alpha_val = 1 / n())

#   Plot the spots aligned to the new Visium grid
p <- ggplot(coords_plotting) +
    geom_point(
        aes(
            x = pxl_col_rounded,
            y = max(pxl_row_rounded) - pxl_row_rounded,
            color = sample_id,
            alpha = alpha_val
        ),
        size = 0.1
    ) +
    coord_fixed() +
    labs(
        x = "Full-res pixels (width)", y = "Full-res pixels (height)",
        color = "Sample ID", title = "Spot Positions (Aligned to Array)"
    ) +
    guides(
        color = guide_legend(override.aes = list(size = 1.5)),
        alpha = "none"
    )

#   'Br8492' will be used for a manuscript figure, where we have stricter
#   aesthetic preferences tailored to this donor
if (opt$donor == "Br8492") {
    p = p + scale_color_manual(values = sample_colors)
}

pdf(file.path(plot_dir, "aligned_spots.pdf"))
print(p)
dev.off()

#-------------------------------------------------------------------------------
#   Plots measuring mapping of multiple spots to the same array coordinates
#-------------------------------------------------------------------------------

#   Verify that for a single sample, mappings were unique
dup_coords <- coords |>
    group_by(sample_id, array_col, array_row) |>
    summarize(n = n()) |>
    ungroup()

p1 <- ggplot(dup_coords) +
    geom_histogram(
        aes(x = n, fill = sample_id),
        bins = length(unique(dup_coords$n)),
        color = "black"
    ) +
    facet_wrap(~sample_id, nrow = 1) +
    labs(
        x = "Num spots per array coordinate", fill = "Sample ID",
        title = "Within-sample duplication of spot mappings"
    )

#   Then check how many spots mapped to the same array coordinates, agnostic
#   to sample
dup_coords <- coords |>
    group_by(array_col, array_row) |>
    summarize(n = n()) |>
    ungroup()

p2 <- ggplot(dup_coords) +
    geom_histogram(aes(x = n), bins = length(unique(dup_coords$n))) +
    labs(
        x = "Num spots per array coordinate",
        title = "Overall duplication of spot mappings"
    )

pdf(file.path(plot_dir, "spot_duplication.pdf"))
plot_grid(p1, p2, nrow = 2)
dev.off()

p <- coords |>
    #   Clean tibble into a nice format for ggplot
    pivot_longer(
        cols = c(
            pxl_row_in_fullres, pxl_row_rounded, pxl_col_in_fullres,
            pxl_col_rounded
        ),
        values_to = "pxl_coord",
        names_pattern = "^(pxl_col|pxl_row)_(in_fullres|rounded)$",
        names_to = c("axis", "position")
    ) |>
    pivot_wider(names_from = axis, values_from = pxl_coord) |>
    mutate(
        #   Use better names
        position = case_when(
            position == "in_fullres" ~ "original",
            position == "rounded" ~ "aligned",
            TRUE ~ position
        )
    ) |>
    #   Plot the original spot coordiates on top of their alignments,
    #   faceted by sample
    ggplot() +
    geom_point(
        aes(x = pxl_col, y = max(pxl_row) - pxl_row, color = position),
        size = 0.1
    ) +
    facet_wrap(~sample_id) +
    coord_fixed()

pdf(file.path(plot_dir, "alignment_by_sample.pdf"), width = 14, height = 14)
print(p)
dev.off()

################################################################################
#   Measure error in aligning (rounding) pixel coordinates to fit array
################################################################################

#   Mean alignment error in spot diameters
print("Mean alignment error in spot diameters:")
coords |>
    summarize(
        d = mean(
            (pxl_row_in_fullres - pxl_row_rounded)**2 +
                (pxl_col_in_fullres - pxl_col_rounded)**2
        )**0.5 / sr_json$spot_diameter_fullres
    ) |>
    pull(d) |>
    round(2) |>
    print()

#-------------------------------------------------------------------------------
#   Explore why/ where duplication occurs
#-------------------------------------------------------------------------------

#   Take only duplicated mappings
dup_df_orig <- coords |>
    group_by(array_row, array_col, sample_id) |>
    mutate(num_members = n()) |>
    filter(num_members > 1)

#   If duplicated sites exist, plot some examples
if (nrow(dup_df_orig) > 0) {
    #   Take 4 random duplicated mappings from each sample ID
    dup_df_orig <- dup_df_orig |>
        slice_head(n = 1) |>
        group_by(sample_id) |>
        sample_n(size = 4, replace = TRUE) |>
        select(!num_members)

    plot_list <- list()
    for (i in 1:nrow(dup_df_orig)) {
        #   Grab the set of source spots mapping to the same destination spot,
        #   mark them as duplicates, and use their original pixel coordinates
        dup_df <- coords |>
            filter(
                sample_id == dup_df_orig$sample_id[i],
                array_row == dup_df_orig$array_row[i],
                array_col == dup_df_orig$array_col[i]
            ) |>
            mutate(
                coord_type = "duplicate",
                pxl_row = pxl_row_in_fullres,
                pxl_col = pxl_col_in_fullres
            )

        #   Grab the set of nearby source spots surrounding the duplicate
        #   mapping, and use the rounded pixel coordinates from them
        dup_df <- coords |>
            filter(
                sample_id == dup_df_orig$sample_id[i],
                abs(array_row - dup_df_orig$array_row[i]) <= 4,
                abs(array_col - dup_df_orig$array_col[i]) <= 7,
            ) |>
            mutate(
                coord_type = "nearby",
                pxl_row = pxl_row_rounded,
                pxl_col = pxl_col_rounded
            ) |>
            rbind(dup_df)

        plot_list[[i]] <- ggplot(dup_df) +
            geom_point(aes(x = pxl_col, y = pxl_row, color = coord_type)) +
            coord_fixed()

        #   Grab the legend from the first plot (which matches all other
        #   legends)
        if (i == 1) {
            legend <- get_legend(plot_list[[i]])
        }

        plot_list[[i]] <- plot_list[[i]] +
            theme(
                axis.text.x = element_blank(),
                axis.text.y = element_blank(),
                axis.title.x = element_blank(),
                axis.ticks.x = element_blank(),
                axis.ticks.y = element_blank(),
                axis.title.y = element_blank(),
                legend.position = "none"
            )
    }

    plot_list[["legend"]] <- legend
    pdf(file.path(plot_dir, "duplicated_sites.pdf"))
    grid.arrange(
        grobs = plot_list, bottom = "pixel col", left = "pixel row", ncol = 4
    )
    dev.off()
}

session_info()
